requirements.txt
fastapi
uvicorn
sentence-transformers
faiss-cpu
pydantic
pydantic-settings
pytest
requests
openai